pyproject.toml
[project]
name = "rerun-mjcf-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dm-control>=1.0.30",
    "mujoco>=3.3.2",
    "natsort>=8.4.0",
    "pybind11-stubgen>=2.5.3",
    "pyglet<2",
    "rerun-sdk>=0.23.1",
    "trimesh>=4.6.8",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
